refactor: delegate waypoint commands and wait ticks to TrainActionManager
