+++ src/main/java/letrain/mvp/impl/Model.java
@@ -259,7 +259,7 @@ public class Model implements letrain.mvp.Model {
             for (Locomotive loco : locomotives) {
                 Train train = loco.getTrain();
                 if (train != null && train.isAutoMode()) {
-                    train.checkWaypointArrival();
+                    train.actionManager.checkWaypointArrival();
                     train.actionManager.acquireInitialLocks();
                 }
             }
